[WIP] renaming - network
